Improvement/total stats

+++ features/qrscanner/src/main/java/com/chelz/features/qrscanner/domain/entity/ItemEntity.kt
@@ -9,6 +9,6 @@ data class ItemsEntity(
 data class SingleItemEntity(
 	var sum: Int? = null,
 	var name: String? = null,
-	var price: Double? = null,
-	var quantity: Int? = null,
+	var price: Int? = null,
+	var quantity: Double? = null,
 ) : Serializable
