+++ features/qrscanner/src/main/java/com/chelz/features/qrscanner/presentation/result/adapter/QrItem.kt
@@ -5,7 +5,7 @@ import com.chelz.shared.accounts.domain.entity.Category
 data class QrItem(
 	var sum: Int? = null,
 	var name: String? = null,
-	var price: Double? = null,
-	var quantity: Int? = null,
+	var price: Int? = null,
+	var quantity: Double? = null,
 	var category: Category? = null,
 )
